refactoring

+++ TorrentClient/Utils/BigEndian.cs
@@ -17,7 +17,7 @@ namespace TorrentClient.Utils
       return bytes[3] | bytes[2] << 8 | bytes[1] << 16 | bytes[0] << 24;
     }
 
-    public static void PutUint32(Span<byte> bytes, int length)
+    public static void PutUint32(Span<byte> bytes, long length)
     {
       bytes[0] = (byte)(length >> 24);
       bytes[1] = (byte)(length >> 16);
